Face-Similarity-develop/requirements.txt
absl-py==0.9.0
apturl==0.5.2
astor==0.8.1
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
Click==7.0
cmake==3.13.3
command-not-found==0.3
cryptography==1.2.3
dateparser==0.7.2
defer==1.0.6
dlib==19.9.0
face-recognition==1.2.3
face-recognition-models==0.3.0
feedparser==5.1.3
Flask==1.1.1
gast==0.3.3
grpcio==1.27.1
guacamole==0.9.2
h5py==2.10.0
html5lib==0.999
httplib2==0.9.1
idna==2.0
impyute==0.0.8
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface==0.6
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.2.1
MarkupSafe==1.1.1
mtcnn==0.1.0
numpy==1.18.1
oauthlib==1.0.3
onboard==1.2.0
opencv-python==4.2.0.32
padme==1.1.1
pexpect==4.0.1
Pillow==7.0.0
plainbox==0.25
protobuf==3.11.3
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pydot==1.4.1
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.4.6
python-apt==1.1.0b1+ubuntu0.16.4.8
python-dateutil==2.8.1
python-debian==0.1.27
python-systemd==231
pytz==2019.3
pyxdg==0.25
PyYAML==5.3
regex==2020.1.8
reportlab==3.3.0
requests==2.9.1
scikit-learn==0.22.1
scipy==1.4.1
sessioninstaller==0.0.0
six==1.14.0
system-service==0.3
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
tzlocal==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
Werkzeug==1.0.0
xdiagnose==3.8.4.1
xgboost==0.90
xkit==0.0.0
XlsxWriter==0.7.3
